Move webhook test utilities to main TestCase class
